[todo]: added todo API resource

+++ app/Http/Controllers/TodoController.php
@@ -4,26 +4,43 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\TodoRequest;
 use App\Models\Todo;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Spatie\QueryBuilder\QueryBuilder;
 
 class TodoController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        return Todo::all();
+
+        $todos = QueryBuilder::for(Todo::class)
+            ->where('user_id', auth()->user()->id)
+            ->allowedFilters('title', 'body', 'status')
+            ->allowedSorts('title', 'body', 'status')
+            ->paginate($request->per_page ?? 10);
+
+        return response()->json($todos);
+
     }
 
     public function store(TodoRequest $request)
     {
-        return Todo::create($request->validated());
+        return Auth::user()->todos()->create($request->validated());
     }
 
     public function show(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
         return $todo;
     }
 
     public function update(TodoRequest $request, Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
         $todo->update($request->validated());
 
         return $todo;
@@ -31,6 +48,9 @@ class TodoController extends Controller
 
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
         $todo->delete();
 
         return response()->json();
